applied clippy

+++ crates/libera-reader-core/src/db/models/books/book_hashes.rs
@@ -37,12 +37,9 @@ impl BookHashes {
         let mut updated_self = old_self.clone();
         match updated_self.books.swap_take(book_path) {
           Some(mut old_book_path) => {
-            match &old_book_path.deleted {
-              true => {
-                old_book_path.deleted = false;
-                updated_self.books.insert(old_book_path);
-              }
-              false => {}
+            if old_book_path.deleted {
+              old_book_path.deleted = false;
+              updated_self.books.insert(old_book_path);
             };
           }
           None => {
@@ -60,57 +57,39 @@ impl BookHashes {
     Ok(())
   }
   pub(crate) fn remove_book(book_hash: BookHash, book_path: &BookPath, db: &DB) -> anyhow::Result<()> {
-    match Self::get_by_hash(book_hash, db)? {
-      Some(old_self) => {
-        let mut updated_self = old_self.clone();
-        match updated_self.books.swap_remove(book_path) {
+    if let Some(old_self) = Self::get_by_hash(book_hash, db)? {
+      let mut updated_self = old_self.clone();
+      if updated_self.books.swap_remove(book_path) {
+        match updated_self.books.is_empty() {
           true => {
-            match updated_self.books.is_empty() {
-              true => {
-                db.remove(old_self)?;
-              }
-              false => {
-                db.update(old_self, updated_self)?;
-              }
-            };
+            db.remove(old_self)?;
           }
-          false => {}
-        }
+          false => {
+            db.update(old_self, updated_self)?;
+          }
+        };
       }
-      None => {}
     };
     Ok(())
   }
   pub(crate) fn mark_book_as_deleted(book_hash: BookHash, book_path: &BookPath, db: &DB) -> anyhow::Result<()> {
-    match Self::get_by_hash(book_hash, db)? {
-      Some(old_self) => {
-        let mut updated_self = old_self.clone();
-        match updated_self.books.swap_take(book_path) {
-          Some(mut inn_book_path) => {
-            inn_book_path.mark_as_deleted();
-            updated_self.books.insert(inn_book_path);
-            db.update(old_self, updated_self)?;
-          }
-          None => {}
-        };
-      }
-      None => {}
+    if let Some(old_self) = Self::get_by_hash(book_hash, db)? {
+      let mut updated_self = old_self.clone();
+      if let Some(mut inn_book_path) = updated_self.books.swap_take(book_path) {
+        inn_book_path.mark_as_deleted();
+        updated_self.books.insert(inn_book_path);
+        db.update(old_self, updated_self)?;
+      };
     };
     Ok(())
   }
   pub(crate) fn update_book_path(book_hash: BookHash, old_book_path: &BookPath, new_book_path: &BookPath, db: &DB) -> anyhow::Result<()> {
-    match Self::get_by_hash(book_hash, db)? {
-      Some(old_self) => {
-        let mut updated_self = old_self.clone();
-        match updated_self.books.swap_remove(old_book_path) {
-          true => {
-            updated_self.books.insert(new_book_path.clone());
-            db.update(old_self, updated_self)?;
-          }
-          false => {}
-        };
-      }
-      None => {}
+    if let Some(old_self) = Self::get_by_hash(book_hash, db)? {
+      let mut updated_self = old_self.clone();
+      if updated_self.books.swap_remove(old_book_path) {
+        updated_self.books.insert(new_book_path.clone());
+        db.update(old_self, updated_self)?;
+      };
     };
     Ok(())
   }
